feat: v2 platform sdk

+++ lib/Model/ViewNumberOfChannelsWithUnreadMessagesResponse.php
@@ -13,7 +13,7 @@
 /**
  * Sendbird Platform SDK
  *
- * Sendbird Platform API SDK  https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api
+ * Sendbird Platform API SDK  [https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api](https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api)  Contact Support:   Name: Sendbird   Email: [support@sendbird.com](https://mailto:support@sendbird.com)
  *
  * The version of the OpenAPI document: 1.0.0
  * Contact: support@sendbird.com
@@ -60,7 +60,7 @@ class ViewNumberOfChannelsWithUnreadMessagesResponse implements ModelInterface, 
       * @var string[]
       */
     protected static $openAPITypes = [
-        'unread_count' => 'float'
+        'unread_count' => 'int'
     ];
 
     /**
@@ -209,7 +209,7 @@ class ViewNumberOfChannelsWithUnreadMessagesResponse implements ModelInterface, 
     /**
      * Gets unread_count
      *
-     * @return float|null
+     * @return int|null
      */
     public function getUnreadCount()
     {
@@ -219,7 +219,7 @@ class ViewNumberOfChannelsWithUnreadMessagesResponse implements ModelInterface, 
     /**
      * Sets unread_count
      *
-     * @param float|null $unread_count unread_count
+     * @param int|null $unread_count unread_count
      *
      * @return self
      */
