+++ lib/Model/ViewNumberOfMonthlyActiveUsersResponse.php
@@ -13,7 +13,7 @@
 /**
  * Sendbird Platform SDK
  *
- * Sendbird Platform API SDK  https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api
+ * Sendbird Platform API SDK  [https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api](https://sendbird.com/docs/chat/v3/platform-api/getting-started/prepare-to-use-api)  Contact Support:   Name: Sendbird   Email: [support@sendbird.com](https://mailto:support@sendbird.com)
  *
  * The version of the OpenAPI document: 1.0.0
  * Contact: support@sendbird.com
@@ -60,7 +60,7 @@ class ViewNumberOfMonthlyActiveUsersResponse implements ModelInterface, ArrayAcc
       * @var string[]
       */
     protected static $openAPITypes = [
-        'mau' => 'float'
+        'mau' => 'int'
     ];
 
     /**
@@ -209,7 +209,7 @@ class ViewNumberOfMonthlyActiveUsersResponse implements ModelInterface, ArrayAcc
     /**
      * Gets mau
      *
-     * @return float|null
+     * @return int|null
      */
     public function getMau()
     {
@@ -219,7 +219,7 @@ class ViewNumberOfMonthlyActiveUsersResponse implements ModelInterface, ArrayAcc
     /**
      * Sets mau
      *
-     * @param float|null $mau mau
+     * @param int|null $mau mau
      *
      * @return self
      */
